(subscription) Do not update or delete a subscription that is not current
This because Stripe creates a subscription object also if the payment fails, and sets it to expire within 24 hours if a payment is not made. However, if a user starts a new checkout session, and succeeds in creating a subscription, that's the subscription that will be deleted in 24 hours! Stripe supports multiple user subscriptions of the same kind, while we don't.

+++ astrobin_apps_payments/services/stripe_webhook_service.py
@@ -188,6 +188,14 @@ class StripeWebhookService(object):
             UserSubscription, user=user, subscription=subscription
         )
 
+        if user.userprofile.stripe_subscription_id != session['id']:
+            log.info(
+                "stripe_webhook: user %s subscription id %s does not match session id %s" % (
+                    user.pk, user.userprofile.stripe_subscription_id, session['id']
+                )
+            )
+            return
+
         if user_subscription:
             user_subscription.delete()
 
@@ -203,6 +211,14 @@ class StripeWebhookService(object):
         user_subscription: UserSubscription = get_object_or_None(
             UserSubscription, user=user, subscription=subscription
         )
+
+        if user.userprofile.stripe_subscription_id != session['id']:
+            log.info(
+                "stripe_webhook: user %s subscription id %s does not match session id %s" % (
+                    user.pk, user.userprofile.stripe_subscription_id, session['id']
+                )
+            )
+            return
 
         if user_subscription is None:
             user_subscription = UserSubscription.objects.create(
